+++ nani/tests/__init__.py
@@ -1,10 +1,11 @@
 from nani.tests.admin import (NormalAdminTests, AdminEditTests, 
     AdminNoFixturesTests)
 from nani.tests.basic import (OptionsTest, BasicQueryTest, CreateTest, GetTest, 
-    TranslatedTest, DeleteLanguageCodeTest, GetByLanguageTest, FallbackTest, 
-    DescriptorTests, DefinitionTests)
+    TranslatedTest, DeleteLanguageCodeTest, GetByLanguageTest, DescriptorTests, 
+    DefinitionTests)
 from nani.tests.dates import LatestTests, DatesTests
 from nani.tests.docs import DocumentationTests
+from nani.tests.fallbacks import FallbackTests
 from nani.tests.fieldtranslator import FieldtranslatorTests
 from nani.tests.forms import FormTests
 from nani.tests.ordering import OrderingTest