+++ nani/tests/dates.py
@@ -1,11 +1,10 @@
 from nani.test_utils.data import DATES_REVERSED, D3
+from nani.test_utils.fixtures import DatesMixin
 from nani.test_utils.testcase import NaniTestCase
 from testproject.app.models import Date
 import datetime
 
-class LatestTests(NaniTestCase):
-    fixtures = ['dates.json']
-    
+class LatestTests(NaniTestCase, DatesMixin):
     def test_shared_latest(self):
         latest = Date.objects.language('en').latest('shared_date')
         self.assertEqual(latest.pk, DATES_REVERSED[D3]['shared_date'])
@@ -16,9 +15,7 @@ class LatestTests(NaniTestCase):
         latest = Date.objects.language('ja').latest('translated_date')
         self.assertEqual(latest.pk, DATES_REVERSED[D3]['translated_date_ja'])
 
-class DatesTests(NaniTestCase):
-    fixtures = ['dates.json']
-    
+class DatesTests(NaniTestCase, DatesMixin):
     def test_objects_dates(self):
         d2011 = datetime.date(year=2011, month=1, day=1)
         self.assertEqual(len(Date.objects.language('en').dates("shared_date", "year")), 2)